refactor(web): split 5 largest Vue components (Closes #287)
