+++ web/src/views/app-detail/RollbackDialog.vue
@@ -2,29 +2,29 @@
 import { useI18n } from 'vue-i18n'
 import AlertDialog from '@/components/ui/AlertDialog.vue'
 
-const { t } = useI18n()
-
-const props = defineProps<{
+defineProps<{
   open: boolean
   appName: string
   currentVersion: string
 }>()
 
-const emit = defineEmits<{
-  'update:open': [value: boolean]
-  confirm: []
+defineEmits<{
+  (e: 'update:open', value: boolean): void
+  (e: 'confirm'): void
 }>()
+
+const { t } = useI18n()
 </script>
 
 <template>
   <AlertDialog
-    :open="open"
+    :model-value="open"
     :title="t('appDetail.rollbackConfirm')"
     :description="t('appDetail.rollbackConfirmDesc', { name: appName, version: currentVersion })"
     :confirm-text="t('appDetail.confirmRollback')"
     :cancel-text="t('common.cancel')"
     variant="destructive"
-    @update:open="emit('update:open', $event)"
-    @confirm="emit('confirm')"
+    @update:model-value="$emit('update:open', $event)"
+    @confirm="$emit('confirm')"
   />
 </template>
